% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_in_n.R
\name{cut_in_n}
\alias{cut_in_n}
\title{Cut a numeric vector into n equal-sized groups}
\usage{
cut_in_n(x, n = 4, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer specifying the number of groups.}

\item{na.rm}{A logical value indicating whether NA values should be removed.}
}
\description{
This function applies base R's \code{cut} and \code{quantile} functions to
cut a numeric vector into n equal-sized groups. It behaves similarly
to the \code{cut} function, with added functionality.
}
\examples{

data(iris)
cut_in_n(iris$Sepal.Length, n = 3)

}
\seealso{
\code{\link{cut}}, \code{\link{quantile}}.
}
