% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dataentryserver}
\alias{dataentryserver}
\title{The server function of shinySQLite}
\usage{
dataentryserver(input, output, session, authenticate = F)
}
\arguments{
\item{input}{Shiny input object passed from the server.}

\item{output}{Shiny output object passed from the server.}

\item{session}{Shiny session object passed from the server.
See \link{https://github.com/nasilabednego/shinySQLite} for more information.}
}
\description{
This is function a server function that renders the outputs as well as sending statements to SQLite database.
}
\details{
This function is used together with it's ui companion,  dataentryui(mytheme)
}
\examples{
library(shiny)
library(RSQLite)
library(DT)
library(shinySQLite)
ui <- dataentryui(title='myapp',analysistab='tab')
server<-function(input,output,session){
dataentryserver(input = input,output = output,session = session)
}
shinyApp(ui, server)

}
