% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dataentryui}
\alias{dataentryui}
\title{The ui function of shinySQLite}
\usage{
dataentryui(title, analysistab)
}
\arguments{
\item{title}{Add title to your ui.}

\item{analysistab}{Add your analysis bs4Dash:::tabItem to existing shinySQLite bs4Dash:::tabItems. This enable you to apply active shinySQLite tables to your data analysis. e.g in server.R you can output$table <- renderTable({useactivedata(input, session)}) and read the table in your analysis bs4Dash:::tabItem e.g tableOutput(dataframe)
See \link{https://github.com/nasilabednego/shinySQLite} for more information.}
}
\description{
This is function a ui function that displays your shinySQLite databases, data frames and dynamic input controls.
}
\details{
This function is used together with it's server companion,  dataentryserver(input, output, session)
}
\examples{
library(shiny)
library(RSQLite)
library(DT)
library(shinySQLite)
ui <- dataentryui(title = 'Kazi database')
server<-function(input,output,session){
   dataentryserver(input = input,output = output,session = session)
}
shinyApp(ui, server)

}
