% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runExample}
\alias{runExample}
\title{Run example shinySQLite app}
\usage{
runExample(example)
}
\description{
Run example shinySQLite app
}
\details{
Quickly launch example app
}
\examples{
runExample(example)

       See \link{https://github.com/nasilabednego/shinySQLite} for more information.
}
