% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{useactivedata}
\alias{useactivedata}
\title{Use your opened data frame from shinySQLite}
\usage{
useactivedata(input, session)
}
\arguments{
\item{input}{Shiny input object passed from the server.
See \link{https://github.com/nasilabednego/shinySQLite} for more information.}
}
\description{
You can use shinySQLite data frame in your functions by typing useactivedata(input)
}
\details{
You can use useactivedata(input) as your data.frame in functions such as renderPlot, renderPrint ... etc.
}
\examples{
library(shiny)
library(RSQLite)
library(DT)
library(shinySQLite)
ui <- dataentryui(title = 'Biashara Group LTD',
          analysistab = bs4Dash:::tabItem(id = 'myAnalysis', tableOutput(dataframe)))

server<-function(input,output,session){
   dataentryserver(input = input,output = output,session = session)
   output$table <- renderTable({
      data <- useactivedata(input = input, session = session)
      data
   })
}
shinyApp(ui, server)

}
