% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save Data Table to Various Formats with Logging}
\usage{
save_data(
  dt,
  key,
  outfile,
  logfile = NULL,
  replacelog = TRUE,
  mask_vars = NULL,
  compress_opt = 50
)
}
\arguments{
\item{dt}{A data table.}

\item{key}{A key that identifies the unique combination of rows.}

\item{outfile}{The name of the output file.}

\item{logfile}{The name of the log file. If NULL, a log file with name \code{data_file_manifest.log} is created in the same directory as \code{outfile}.
Deafults to NULL. Set to FALSE to not create a log file.}

\item{replacelog}{A boolean flag to control whether the generated log file replaces an existing logfile or not. Deafults to TRUE.}

\item{mask_vars}{A vector of names of numeric variables to be excluded from the summary in the log file. Defaults to NULL.}

\item{compress_opt}{The compression level for fst files. Defaults to 50.}
}
\value{
A message indicating successful saving of the data file.
}
\description{
This function saves a data table to various formats including .csv, .dta, .fst, .feather, and .RDS.
It also generates a log file containing the summary of a given data table by calling the \code{generate_log_file} function.
}
\examples{
data(iris)
save_data(iris, key = "Species", outfile = "iris.csv")

}
