% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sd.R
\name{weighted_sd}
\alias{weighted_sd}
\title{Weighted Standard Deviation}
\usage{
weighted_sd(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector of weights.}

\item{na.rm}{A logical value indicating whether NA values in both \code{x} and \code{w} should be removed.}
}
\value{
The weighted standard deviation.
}
\description{
This function computes the weighted standard deviation of a numeric vector.
It leverages the \code{weighted_var} function to compute the variance and then
takes the square root.
}
\examples{

x <- c(1, 2, 3, 4, 5)
w <- c(0.1, 0.2, 0.3, 0.2, 0.2)
weighted_sd(x, w)

}
\seealso{
\code{\link{sd}}, \code{\link{var}}, \code{\link{mean}}, \code{\link{weighted.mean}}, \code{\link{weighted_var}}.
}
